requirements.txt
python-telegram-bot[job-queue]==20.7
openai==1.12.0
python-dotenv==1.1.0
pytz==2024.1
imghdr==0.0.1
